Kaggle MedGemma Impact Challenge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cxr-sentinel"
version = "1.5.0"
description = "Multi-Agent Adversarial Framework for Clinical AI Validation"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Francesco Orsi"}
]
keywords = [
    "medical-ai", "radiology", "chest-x-ray", "adversarial-validation",
    "medgemma", "txgemma", "clinical-ai", "hallucination-detection",
    "multi-agent", "fda-samd"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "bitsandbytes>=0.43.0",
    "peft>=0.10.0",
    "accelerate>=0.30.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "scipy>=1.11.0",
    "rouge-score>=0.1.2",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.urls]
Homepage = "https://github.com/francescoorsi/cxr-sentinel"
Documentation = "https://github.com/francescoorsi/cxr-sentinel/tree/main/docs"
Repository = "https://github.com/francescoorsi/cxr-sentinel"
Issues = "https://github.com/francescoorsi/cxr-sentinel/issues"

[project.scripts]
cxr-sentinel = "scripts.run_pipeline:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
